feat(amazonq): add reference log for chat messages
