+++ packages/amazonq/src/lsp/chat/messages.ts
@@ -57,13 +57,14 @@ import * as vscode from 'vscode'
 import { Disposable, LanguageClient, Position, TextDocumentIdentifier } from 'vscode-languageclient'
 import * as jose from 'jose'
 import { AmazonQChatViewProvider } from './webviewProvider'
-import { AuthUtil } from 'aws-core-vscode/codewhisperer'
+import { AuthUtil, ReferenceLogViewProvider } from 'aws-core-vscode/codewhisperer'
 import { amazonQDiffScheme, AmazonQPromptSettings, messages, openUrl } from 'aws-core-vscode/shared'
 import {
     DefaultAmazonQAppInitContext,
     messageDispatcher,
     EditorContentController,
     ViewDiffMessage,
+    referenceLogText,
 } from 'aws-core-vscode/amazonq'
 import { telemetry, TelemetryBase } from 'aws-core-vscode/telemetry'
 import { isValidResponseError } from './error'
@@ -531,13 +532,17 @@ async function handlePartialResult<T extends ChatResult>(
             tabId: tabId,
         })
     }
+
+    for (const ref of decryptedMessage.codeReference ?? []) {
+        ReferenceLogViewProvider.instance.addReferenceLog(referenceLogText(ref))
+    }
 }
 
 /**
  * Decodes the final chat responses from the language server before sending it to mynah UI.
  * Once this is called the answer response is finished
  */
-async function handleCompleteResult<T>(
+async function handleCompleteResult<T extends ChatResult>(
     result: string | T,
     encryptionKey: Buffer | undefined,
     provider: AmazonQChatViewProvider,
@@ -545,12 +550,15 @@ async function handleCompleteResult<T>(
     disposable: Disposable
 ) {
     const decryptedMessage =
-        typeof result === 'string' && encryptionKey ? await decodeRequest(result, encryptionKey) : result
+        typeof result === 'string' && encryptionKey ? await decodeRequest<T>(result, encryptionKey) : (result as T)
     void provider.webview?.postMessage({
         command: chatRequestType.method,
         params: decryptedMessage,
         tabId: tabId,
     })
+    for (const ref of decryptedMessage.codeReference ?? []) {
+        ReferenceLogViewProvider.instance.addReferenceLog(referenceLogText(ref))
+    }
     disposable.dispose()
 }
