+++ packages/core/src/amazonqFeatureDev/constants.ts
@@ -26,14 +26,6 @@ export const clientErrorMessages = [
     'The folder you chose did not contain any source files in a supported language. Choose another folder and try again.',
 ]
 
-// License text that's used in codewhisperer reference log
-export const referenceLogText = (reference: CodeReference) =>
-    `[${new Date().toLocaleString()}] Accepted recommendation from Amazon Q. Code provided with reference under <a href="${LicenseUtil.getLicenseHtml(
-        reference.licenseName
-    )}" target="_blank">${reference.licenseName}</a> license from repository <a href="${
-        reference.url
-    }" target="_blank">${reference.repository}</a>.<br><br>`
-
 // License text that's used in the file view
 export const licenseText = (reference: CodeReference) =>
     `<a href="${LicenseUtil.getLicenseHtml(reference.licenseName)}" target="_blank">${
